Extract the garage vehicle record factory
bootstrap._selected_vehicle builds every owned vehicle at startup. Buying a
vehicle has to build one more at runtime, with the identical crew generation,
module catalogue, ammunition closure and record shape. Two copies of that would
drift, so the per-vehicle body and the pure helpers it needs move to
vehicle_records, which touches no BigWorld and takes the already imported items
modules as arguments.

The factory takes the two distinctions a career needs but the sandbox does not:
top_modules fits the vehicle the way the historical garage does, and
own_researchable_modules owns every module the type can carry. A bought vehicle
will arrive stock and own only what is installed.

Inventory ids stay derived per session rather than persisted. Nothing needs
them to survive a restart: garage_store keys vehicles on their type compact
descriptor precisely because ids renumber, and a battle receipt names a type.
What a purchase needs is an allocator that cannot collide with a live record,
so the factory takes the id from its caller instead of computing one.

No behaviour change.
